Remove beforeEach from tests
Previously codegen was called within beforeEach.
When that fails, jest simply continues to execute all test cases.
The error from beforeEach is not shown and some expect call is going to fail in the test case,
with an error message that doesn't make it obvious where the actual problem lies.